requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
requests>=2.31.0
python-dotenv>=1.0.0
pytest

httpx>=0.24.1
python-multipart>=0.0.6
geopy>=2.4.0
langchain>=0.0.267
python-jose>=3.3.0

langfuse>=2.0.0 